CRUDmoreFRONTEND/backend/requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.32
Flask-Cors==3.0.10